refactor(tauri.js) keep folder structure on /dist
