+++ cli/tauri.js/src/api/build.ts
@@ -1,7 +1,5 @@
 import { TauriConfig } from 'types'
 import merge from 'webpack-merge'
-import * as entry from '../entry'
-import { tauriDir } from '../helpers/app-paths'
 const getTauriConfig = require('../helpers/tauri-config')
 import Runner from '../runner'
 
@@ -17,8 +15,6 @@ module.exports = async (config: TauriConfig): Promise<void> => {
       config as any
     ) as TauriConfig
   )
-
-  entry.generate(tauriDir, tauriConfig)
 
   return tauri.build(tauriConfig)
 }
