+++ cli/tauri.js/src/api/dev.ts
@@ -1,7 +1,5 @@
 import { TauriConfig } from 'types'
 import merge from 'webpack-merge'
-import * as entry from '../entry'
-import { tauriDir } from '../helpers/app-paths'
 const getTauriConfig = require('../helpers/tauri-config')
 import Runner from '../runner'
 
@@ -18,8 +16,6 @@ module.exports = async (config: TauriConfig): Promise<void> => {
       config as any
     ) as TauriConfig
   )
-
-  entry.generate(tauriDir, tauriConfig)
 
   return tauri.run(tauriConfig)
 }
