+++ cli/tauri.js/src/entry.ts
@@ -1,14 +1,11 @@
-import { ensureDirSync, readFileSync, writeFileSync } from 'fs-extra'
+import { ensureDirSync, writeFileSync } from 'fs-extra'
 import compileTemplate from 'lodash.template'
 import path from 'path'
 import { TauriConfig } from './types/config'
 
 export const generate = (outDir: string, cfg: TauriConfig): void => {
   // this MUST be from the templates repo
-  const apiTemplate = readFileSync(
-    path.resolve(__dirname, '../templates/tauri.js'),
-    'utf-8'
-  )
+  const apiTemplate = require('../templates/tauri.js').default
   const apiContent = compileTemplate(apiTemplate)(cfg)
 
   ensureDirSync(outDir)
